+++ src/components/Calender/Week/WeeklyCalender.tsx
@@ -1,4 +1,4 @@
-import { FlexBox } from 'components/Base';
+import { FlexBox, Typography } from 'components/Base';
 import WeekDays from 'components/Calender/Week/WeekDays';
 import { useCalender } from 'components/useCalender';
 
@@ -18,7 +18,11 @@ export default function WeeklyCalender() {
     return (
         <FlexBox gap={20}>
             {getWeeks()[weeklyWeek].map((day) => (
-                <WeekDays>{day}</WeekDays>
+                <WeekDays>
+                    <Typography component="span" customCSS={{ color: day === String(date) ? 'tomato' : '' }}>
+                        {day}
+                    </Typography>
+                </WeekDays>
             ))}
         </FlexBox>
     );
